Refactoring: unifying coordinate system related variables name to end with "Screen, World, Physics". Split components to several *.h files.

+++ src/utils/resources/aseprite_data.h
@@ -38,8 +38,8 @@ struct AsepriteData
     struct FrameTag
     {
         std::string name;
-        int from;
-        int to;
+        size_t from;
+        size_t to;
     };
 
     std::unordered_map<std::string, FrameTag> frameTags;
